+++ shopify_sdk/gql/versions/v2025_10/types/payload.py
@@ -28,6 +28,7 @@ class BulkOperationRunMutationPayload(AutoRegisterModel):
 class BulkOperationResultPayload(AutoRegisterModel):
     data: Optional[Dict] = Field(default={})
     lineNumber: Optional[int] = Field(alias="__lineNumber", default=None)
+    parentId: Optional[ID] = Field(alias="__parentId", default=None)
     errors: Optional[List[Dict]] = Field(default=None)